requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
